+++ docs/release-notes.md
@@ -1,5 +1,9 @@
 # Release notes
 
+## v0.1.0-rc3.2
+
+See [v0.1.0-rc3.2 release notes](release-notes/v0.1.0-rc3.2.md).
+
 ## v0.1.0-rc3.1
 
 See [v0.1.0-rc3.1 release notes](release-notes/v0.1.0-rc3.1.md).